File walk returns file count

+++ src/explore.h
@@ -42,5 +42,5 @@ void anu_fileq_destroy(anuFileQ* q);
 int anu_fileq_enqueue(anuFileQ* q, anuFile* file_in);
 int anu_fileq_dequeue(anuFileQ* q, anuFile* file_out);
 int anu_open_dir(char* dir_path, DIR** out);
-int anu_recursive_filewalk(char* searchp, anuFileQ* files_out);
+size_t anu_recursive_filewalk(char* searchp, anuFileQ* files_out);
 #endif  // EXPLORE_H_
